requirements.txt
# --- Core Backend ---
fastapi==0.95.2
uvicorn[standard]==0.23.2
pydantic==1.10.13
python-dotenv==1.0.1

# --- Async Tools ---
aiohttp==3.8.6
httpx==0.26.0

# --- Kafka + Stream Processing ---
faust-streaming[rocksdb]
kafka-python==2.2.15

# --- Redis (for pub-sub/cache/rate limiting) ---
redis==4.5.5

# --- DB & ORM ---
asyncpg==0.28.0
SQLAlchemy==1.4.52
alembic==1.12.1

# --- Search & Analytics ---
clickhouse-connect==0.6.10
elasticsearch==8.10.0  # Use this or opensearch-py==2.3.1

# --- Auth & Security ---
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# --- Alerting & ML ---
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.23.5

# --- Observability (Optional) ---
prometheus-client==0.16.0
structlog==23.1.0
loguru==0.7.0

# --- Testing ---
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.2

# --- Dev Tools ---
black==23.7.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0
